Modal window when deleting, limiting the number of services, changes on the main page
On the main page, summary statistics for all services have been removed and a column with SLA for a specific service has been added.
When deleting, a modal window has been added to confirm the deletion of the service.
Added a limit on the number of services for one user.